Update CorsFilter, Update login responseBody
