+++ src/main/java/com/example/csws/config/CorsConfig.java
@@ -1,19 +1,24 @@
 package com.example.csws.config;
 
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
+import org.springframework.web.cors.CorsConfiguration;
+import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
+import org.springframework.web.filter.CorsFilter;
 
 @Configuration
-public class CorsConfig implements WebMvcConfigurer {
-   @Override
-   public void addCorsMappings(CorsRegistry registry) {
-      registry.addMapping("/**")
-              .allowedOriginPatterns("*") // 모든 출처 허용. 필요에 따라 특정 출처만 지정할 수 있습니다.
-              .allowedMethods("GET", "POST", "PUT", "DELETE") // 허용할 HTTP 메소드 지정
-              .allowedHeaders("*") // 허용할 요청 헤더 지정
-              .allowCredentials(true) // 자격 증명(쿠키, 인증 헤더 등) 허용 여부 설정
-              .maxAge(3600); // 사전 요청(preflight)의 유효 기간 설정 (초 단위)
+public class CorsConfig {
+
+   @Bean
+   public CorsFilter corsFilter() {
+      UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
+      CorsConfiguration config = new CorsConfiguration();
+      config.setAllowCredentials(true);       // 내 서버가 응답할 때 json을 JS에서 처리할 수 있게 설정
+      config.addAllowedOriginPattern("*");           // 모든 ip에 응답을 허용
+      config.addAllowedHeader("*");           // 모든 header에 응답 허용
+      config.addAllowedMethod("*");           // 모든 post,get,put,delete,patch 요청허용
+      source.registerCorsConfiguration("/**",config);
+      return new CorsFilter(source);
    }
+
 }
